Consolidate driver-interaction code into pkg/minikube/driver
