+++ pkg/minikube/cluster/cluster_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/docker/machine/libmachine/state"
 	"k8s.io/minikube/pkg/minikube/config"
 	"k8s.io/minikube/pkg/minikube/constants"
+	"k8s.io/minikube/pkg/minikube/driver"
 	"k8s.io/minikube/pkg/minikube/registry"
 	"k8s.io/minikube/pkg/minikube/tests"
 )
@@ -43,13 +44,13 @@ func createMockDriverHost(c config.MachineConfig) interface{} {
 
 func RegisterMockDriver(t *testing.T) {
 	t.Helper()
-	_, err := registry.Driver(constants.DriverMock)
+	_, err := registry.Driver(driver.Mock)
 	// Already registered
 	if err == nil {
 		return
 	}
 	err = registry.Register(registry.DriverDef{
-		Name:          constants.DriverMock,
+		Name:          driver.Mock,
 		Builtin:       true,
 		ConfigCreator: createMockDriverHost,
 		DriverCreator: func() drivers.Driver {
@@ -62,7 +63,7 @@ func RegisterMockDriver(t *testing.T) {
 }
 
 var defaultMachineConfig = config.MachineConfig{
-	VMDriver:    constants.DriverMock,
+	VMDriver:    driver.Mock,
 	MinikubeISO: constants.DefaultISOURL,
 	Downloader:  MockDownloader{},
 	DockerEnv:   []string{"MOCK_MAKE_IT_PROVISION=true"},
@@ -216,7 +217,7 @@ func TestStartHostConfig(t *testing.T) {
 	provision.SetDetector(md)
 
 	config := config.MachineConfig{
-		VMDriver:   constants.DriverMock,
+		VMDriver:   driver.Mock,
 		DockerEnv:  []string{"FOO=BAR"},
 		DockerOpt:  []string{"param=value"},
 		Downloader: MockDownloader{},
