+++ pkg/minikube/constants/constants_darwin.go
@@ -20,12 +20,3 @@ package constants
 
 // DefaultMountDir is the default mounting directory for Darwin
 var DefaultMountDir = "/Users"
-
-// SupportedVMDrivers is a list of supported drivers on Darwin.
-var SupportedVMDrivers = [...]string{
-	DriverVirtualbox,
-	DriverParallels,
-	DriverVmwareFusion,
-	DriverHyperkit,
-	DriverVmware,
-}
