+++ pkg/minikube/constants/constants_gendocs.go
@@ -19,15 +19,3 @@ limitations under the License.
 package constants
 
 var DefaultMountDir = "$HOME"
-
-// SupportedVMDrivers is a list of supported drivers on all platforms.
-var SupportedVMDrivers = [...]string{
-	DriverVirtualbox,
-	DriverParallels,
-	DriverVmwareFusion,
-	DriverHyperv,
-	DriverHyperkit,
-	DriverKvm2,
-	DriverVmware,
-	DriverNone,
-}
