+++ pkg/minikube/driver/constants.go
@@ -1,0 +1,14 @@
+package driver
+
+const (
+	// Driver names
+	Mock         = "mock-driver"
+	None         = "none"
+	KVM2         = "kvm2"
+	VirtualBox   = "virtualbox"
+	HyperKit     = "hyperkit"
+	VMware       = "vmware"
+	VMwareFusion = "vmwarefusion"
+	HyperV       = "hyperv"
+	Parallels    = "parallels"
+)
