+++ pkg/minikube/driver/driver.go
@@ -1,0 +1,55 @@
+package driver
+
+import (
+	"fmt"
+	"os"
+)
+
+// SupportedDrivers returns a list of supported drivers
+func SupportedDrivers() []string {
+	return supportedDrivers
+}
+
+// Supported returns if the driver is supported on this host.
+func Supported(name string) bool {
+	for _, d := range supportedDrivers {
+		if name == d {
+			return true
+		}
+	}
+	return false
+}
+
+// BareMetal returns if this driver is unisolated
+func BareMetal(name string) bool {
+	return name == None || name == Mock
+}
+
+// FlagHints are hints for what default options should be used for this driver
+type FlagHints struct {
+	ExtraOptions string
+	CacheImages  bool
+}
+
+// FlagDefaults returns suggested defaults based on a driver
+func FlagDefaults(name string) FlagHints {
+	if name != None {
+		return FlagHints{}
+	}
+
+	// for more info see: https://github.com/kubernetes/minikube/issues/3511
+	f := "/run/systemd/resolve/resolv.conf"
+	extraOpts := ""
+	if _, err := os.Stat(f); err == nil {
+		extraOpts = fmt.Sprintf("kubelet.resolv-conf=%s", f)
+	}
+	return FlagHints{
+		ExtraOptions: extraOpts,
+		CacheImages:  false,
+	}
+}
+
+// Default returns the default driver on this hos
+func Default() string {
+	return VirtualBox
+}
