+++ pkg/minikube/driver/driver_darwin.go
@@ -1,0 +1,20 @@
+package driver
+
+import "os/exec"
+
+// supportedDrivers is a list of supported drivers on Darwin.
+var supportedDrivers = []string{
+	VirtualBox,
+	Parallels,
+	VMwareFusion,
+	HyperKit,
+	VMware,
+}
+
+func VBoxManagePath() string {
+	cmd := "VBoxManage"
+	if path, err := exec.LookPath(cmd); err == nil {
+		return path
+	}
+	return cmd
+}
