+++ pkg/minikube/machine/client.go
@@ -41,7 +41,7 @@ import (
 	"github.com/docker/machine/libmachine/version"
 	"github.com/pkg/errors"
 	"k8s.io/minikube/pkg/minikube/command"
-	"k8s.io/minikube/pkg/minikube/constants"
+	"k8s.io/minikube/pkg/minikube/driver"
 	"k8s.io/minikube/pkg/minikube/exit"
 	"k8s.io/minikube/pkg/minikube/localpath"
 	"k8s.io/minikube/pkg/minikube/out"
@@ -80,27 +80,27 @@ type LocalClient struct {
 }
 
 // NewHost creates a new Host
-func (api *LocalClient) NewHost(driverName string, rawDriver []byte) (*host.Host, error) {
+func (api *LocalClient) NewHost(drvName string, rawDriver []byte) (*host.Host, error) {
 	var def registry.DriverDef
 	var err error
-	if def, err = registry.Driver(driverName); err != nil {
+	if def, err = registry.Driver(drvName); err != nil {
 		return nil, err
 	} else if !def.Builtin || def.DriverCreator == nil {
-		return api.legacyClient.NewHost(driverName, rawDriver)
+		return api.legacyClient.NewHost(drvName, rawDriver)
 	}
 
-	driver := def.DriverCreator()
+	d := def.DriverCreator()
 
-	err = json.Unmarshal(rawDriver, driver)
+	err = json.Unmarshal(rawDriver, d)
 	if err != nil {
 		return nil, errors.Wrapf(err, "Error getting driver %s", string(rawDriver))
 	}
 
 	return &host.Host{
 		ConfigVersion: version.ConfigVersion,
-		Name:          driver.GetMachineName(),
-		Driver:        driver,
-		DriverName:    driver.DriverName(),
+		Name:          d.GetMachineName(),
+		Driver:        d,
+		DriverName:    d.DriverName(),
 		HostOptions: &host.Options{
 			AuthOptions: &auth.Options{
 				CertDir:          api.certsDir,
@@ -148,10 +148,10 @@ func (api *LocalClient) Close() error {
 
 // CommandRunner returns best available command runner for this host
 func CommandRunner(h *host.Host) (command.Runner, error) {
-	if h.DriverName == constants.DriverMock {
+	if h.DriverName == driver.Mock {
 		return &command.FakeCommandRunner{}, nil
 	}
-	if h.DriverName == constants.DriverNone {
+	if driver.BareMetal(h.Driver.DriverName()) {
 		return &command.ExecRunner{}, nil
 	}
 	client, err := sshutil.NewSSHClient(h.Driver)
@@ -194,7 +194,7 @@ func (api *LocalClient) Create(h *host.Host) error {
 		{
 			"waiting",
 			func() error {
-				if h.Driver.DriverName() == constants.DriverNone {
+				if driver.BareMetal(h.Driver.DriverName()) {
 					return nil
 				}
 				return mcnutils.WaitFor(drivers.MachineInState(h.Driver, state.Running))
@@ -203,7 +203,7 @@ func (api *LocalClient) Create(h *host.Host) error {
 		{
 			"provisioning",
 			func() error {
-				if h.Driver.DriverName() == constants.DriverNone {
+				if driver.BareMetal(h.Driver.DriverName()) {
 					return nil
 				}
 				pv := provision.NewBuildrootProvisioner(h.Driver)
@@ -270,11 +270,11 @@ func (cg *CertGenerator) ValidateCertificate(addr string, authOptions *auth.Opti
 	return true, nil
 }
 
-func registerDriver(driverName string) {
-	def, err := registry.Driver(driverName)
+func registerDriver(drvName string) {
+	def, err := registry.Driver(drvName)
 	if err != nil {
 		if err == registry.ErrDriverNotFound {
-			exit.UsageT("unsupported driver: {{.name}}", out.V{"name": driverName})
+			exit.UsageT("unsupported driver: {{.name}}", out.V{"name": drvName})
 		}
 		exit.WithError("error getting driver", err)
 	}
