+++ pkg/minikube/registry/drvs/parallels/doc.go
