+++ pkg/minikube/tunnel/tunnel.go
@@ -29,7 +29,7 @@ import (
 	"github.com/pkg/errors"
 	typed_core "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/minikube/pkg/minikube/config"
-	"k8s.io/minikube/pkg/minikube/constants"
+	"k8s.io/minikube/pkg/minikube/driver"
 )
 
 // tunnel represents the basic API for a tunnel: periodically the state of the tunnel
@@ -150,7 +150,7 @@ func setupRoute(t *tunnel, h *host.Host) {
 			return
 		}
 
-		if h.DriverName == constants.DriverHyperkit {
+		if h.DriverName == driver.HyperKit {
 			// the virtio-net interface acts up with ip tunnels :(
 			setupBridge(t)
 			if t.status.RouteError != nil {
