+++ test/integration/driver_install_or_update_test.go
@@ -84,7 +84,7 @@ func TestKVMDriverInstallOrUpdate(t *testing.T) {
 			t.Fatalf("Expected new semver. test: %v, got: %v", tc.name, err)
 		}
 
-		err = drivers.InstallOrUpdate("kvm2", dir, newerVersion, true)
+		err = driver.InstallOrUpdate("kvm2", dir, newerVersion, true)
 		if err != nil {
 			t.Fatalf("Failed to update driver to %v. test: %s, got: %v", newerVersion, tc.name, err)
 		}
@@ -147,7 +147,7 @@ func TestHyperKitDriverInstallOrUpdate(t *testing.T) {
 			t.Fatalf("Expected new semver. test: %v, got: %v", tc.name, err)
 		}
 
-		err = drivers.InstallOrUpdate("hyperkit", dir, newerVersion, true)
+		err = driver.InstallOrUpdate("hyperkit", dir, newerVersion, true)
 		if err != nil {
 			t.Fatalf("Failed to update driver to %v. test: %s, got: %v", newerVersion, tc.name, err)
 		}
